+++ app/api/serializers/access.py
@@ -1,6 +1,8 @@
 from rest_framework import serializers, request
 from rest_framework.reverse import reverse
-from access.models import Organization, Team
+
+from access.models.organization import Organization
+from access.models.team import Team
 
 from django.contrib.auth.models import Permission
