+++ app/api/tests/unit/test_index_viewset.py
@@ -4,7 +4,7 @@ from django.test import Client, TestCase
 
 from rest_framework.permissions import IsAuthenticated
 
-from access.models import Organization
+from access.models.organization import Organization
 
 from api.tests.abstract.viewsets import ViewSetCommon
