+++ app/api/tests/unit/token/test_token.py
@@ -10,7 +10,9 @@ from django.contrib.auth.models import AnonymousUser, Permission, User
 from django.shortcuts import reverse
 from django.test import TestCase, Client
 
-from access.models import Organization, Team, TeamUsers
+from access.models.organization import Organization
+from access.models.team import Team
+from access.models.team_user import TeamUsers
 
 from api.models.tokens import AuthToken
